+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 (newest entries on top)
 
+## 1.1.0.- 2017-09-17
+- [Issue #1: GPS-Links to other map services](https://github.com/canonknipser/viewexif/issues/1)
+
+	- google maps
+	- open street maps
+
+- [Issue #2: UCP component](https://github.com/canonknipser/viewexif/issues/2)
+
+
+- [Issue #14: Enhanced error message on enable](https://github.com/canonknipser/viewexif/issues/14)
+
+
+- [Issue #16: solve validation issues from 1.0.3](https://github.com/canonknipser/viewexif/issues/16)
+
+	-  config/services.yml
+		- %core.root_path%
+		Line 8: please also put this parameter into quotes, like you already did with the service arguments.
+
+	- event/main_listener.php
+		* @param \phpbb\config	$config		Config object
+		Line 44: you probably meant `\phpbb\config\config `here.
+
+	- Please remove the  large block of commented code, lines 311-359.
+
+- new acp module
+- new ucp module
+- added new language variables in viewexif.php
+- added a few new language files
+
+	- viewexif_acp.php
+	- info_acp_ck_ve.php
+	- info_ucp_ck_ve.php
+
+- new language *de_x_sie* (German formal honorific)
+
 ## 1.0.3-RC - 2017-02-23
 - fixed issue #13 [Load css only for pages containing attachments](https://github.com/canonknipser/viewexif/issues/13)
 - fixed issue #12 ["Division by zero" on calculation GPS coordinates](https://github.com/canonknipser/viewexif/issues/12)